+++ src/utils/SafePermit.sol
@@ -14,15 +14,14 @@ library FastPermit {
         address spender,
         uint256 amount,
         uint256 deadline,
-        uint8 v,
-        bytes32 r,
-        bytes32 s
+        bytes32 vs,
+        bytes32 r
     ) internal returns (bool success) {
         assembly ("memory-safe") {
             let ptr := mload(0x40)
-            mstore(add(0xd4, ptr), s)
-            mstore(add(0xb4, ptr), r)
-            mstore(add(0x94, ptr), and(0xff, v))
+            mstore(add(0xd4, ptr), and(0x7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff, vs)) // `s`.
+            mstore(add(0xb4, ptr), r) // `r`.
+            mstore(add(0x94, ptr), add(0x1b, shr(0xff, vs))) // `v`.
             mstore(add(0x74, ptr), deadline)
             mstore(add(0x54, ptr), amount)
             mstore(add(0x34, ptr), spender)
@@ -41,15 +40,14 @@ library FastPermit {
         uint256 nonce,
         uint256 expiry,
         bool allowed,
-        uint8 v,
-        bytes32 r,
-        bytes32 s
+        bytes32 vs,
+        bytes32 r
     ) internal returns (bool success) {
         assembly ("memory-safe") {
             let ptr := mload(0x40)
-            mstore(add(0xf4, ptr), s)
-            mstore(add(0xd4, ptr), r)
-            mstore(add(0xb4, ptr), and(0xff, v))
+            mstore(add(0xf4, ptr), and(0x7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff, vs)) // `s`.
+            mstore(add(0xd4, ptr), r) // `r`.
+            mstore(add(0xb4, ptr), add(0x1b, shr(0xff, vs))) // `v`.
             mstore(add(0x94, ptr), allowed)
             mstore(add(0x74, ptr), expiry)
             mstore(add(0x54, ptr), nonce)
@@ -67,9 +65,8 @@ library FastPermit {
         address owner,
         address spender,
         uint256 amount,
-        uint8 v,
-        bytes32 r,
-        bytes32 s
+        bytes32 vs,
+        bytes32 r
     ) internal returns (bool success, bytes32 functionSignatureHash) {
         assembly ("memory-safe") {
             let ptr := mload(0x40)
@@ -77,9 +74,9 @@ library FastPermit {
             mstore(add(0xd8, ptr), spender)
             mstore(add(0xc4, ptr), 0x095ea7b3000000000000000000000000) // selector for `approve(address,uint256)` with `spender` padding
             mstore(add(0xb4, ptr), 0x44) // length of approve call
-            mstore(add(0x94, ptr), and(0xff, v))
-            mstore(add(0x74, ptr), s)
-            mstore(add(0x54, ptr), r)
+            mstore(add(0x94, ptr), add(0x1b, shr(0xff, vs))) // `v`.
+            mstore(add(0x74, ptr), and(0x7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff, vs)) // `s`.
+            mstore(add(0x54, ptr), r) // `r`.
             mstore(add(0x34, ptr), 0xa0) // offset to function signature
             mstore(add(0x14, ptr), owner)
             mstore(ptr, 0x0c53c51c000000000000000000000000) // selector for `executeMetaTransaction(address,bytes,bytes32,bytes32,uint8)` with `owner`'s padding
@@ -177,13 +174,13 @@ library SafePermit {
         }
     }
 
-    function _checkSignature(bytes32 signingHash, address owner, uint8 v, bytes32 r, bytes32 s) internal view {
+    function _checkSignature(bytes32 signingHash, address owner, bytes32 vs, bytes32 r) internal view {
         assembly ("memory-safe") {
             let ptr := mload(0x40)
             mstore(0x00, signingHash)
-            mstore(0x20, and(0xff, v))
-            mstore(0x40, r)
-            mstore(0x60, s)
+            mstore(0x20, add(0x1b, shr(0xff, vs))) // `v`.
+            mstore(0x40, r) // `r`.
+            mstore(0x60, and(0x7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff, vs)) // `s`.
             let recovered := mload(staticcall(gas(), 0x01, 0x00, 0x80, 0x01, 0x20))
             if lt(returndatasize(), shl(0x60, xor(owner, recovered))) {
                 mstore(0x00, 0x8baa579f) // selector for `InvalidSignature()`
@@ -200,15 +197,14 @@ library SafePermit {
         address spender,
         uint256 amount,
         uint256 deadline,
-        uint8 v,
-        bytes32 r,
-        bytes32 s
+        bytes32 vs,
+        bytes32 r
     ) internal {
         // `permit` could succeed vacuously with no returndata if there's a fallback
         // function (e.g. WETH). `permit` could fail spuriously if it was
         // replayed/frontrun. Avoid these by manually verifying the effects and
         // signature. Insufficient gas griefing is defused by checking the effects.
-        if (!token.fastPermit(owner, spender, amount, deadline, v, r, s)) {
+        if (!token.fastPermit(owner, spender, amount, deadline, vs, r)) {
             // Check effects and signature
             if (block.timestamp > deadline) {
                 _revert(0x1a15a3cc); // selector for `PermitExpired()`
@@ -234,7 +230,7 @@ library SafePermit {
                 mstore(add(0x40, ptr), keccak256(add(0x40, ptr), 0xc0))
                 signingHash := keccak256(add(0x1e, ptr), 0x42)
             }
-            _checkSignature(signingHash, owner, v, r, s);
+            _checkSignature(signingHash, owner, vs, r);
         }
     }
 
@@ -245,12 +241,11 @@ library SafePermit {
         uint256 nonce,
         uint256 deadline,
         bool allowed,
-        uint8 v,
-        bytes32 r,
-        bytes32 s
+        bytes32 vs,
+        bytes32 r
     ) internal {
         // See comments above
-        if (!token.fastDAIPermit(owner, spender, nonce, deadline, allowed, v, r, s)) {
+        if (!token.fastDAIPermit(owner, spender, nonce, deadline, allowed, vs, r)) {
             // Check effects and signature
             // https://etherscan.io/token/0x6b175474e89094c44da98b954eedeac495271d0f#code#L188
             if ((block.timestamp > deadline).and(deadline > 0)) {
@@ -281,7 +276,7 @@ library SafePermit {
                 mstore(add(0x40, ptr), keccak256(add(0x40, ptr), 0xc0))
                 signingHash := keccak256(add(0x1e, ptr), 0x42)
             }
-            _checkSignature(signingHash, owner, v, r, s);
+            _checkSignature(signingHash, owner, vs, r);
         }
     }
 
@@ -290,13 +285,11 @@ library SafePermit {
         address owner,
         address spender,
         uint256 amount,
-        uint8 v,
-        bytes32 r,
-        bytes32 s
+        bytes32 vs,
+        bytes32 r
     ) internal {
         // See comments above
-        (bool success, bytes32 functionSignatureHash) =
-            token.fastApproveMetaTransaction(owner, spender, amount, v, r, s);
+        (bool success, bytes32 functionSignatureHash) = token.fastApproveMetaTransaction(owner, spender, amount, vs, r);
         if (!success) {
             // Check effects and signature
             uint256 nonce = token.fastNonce(owner, uint32(token.getNonce.selector));
@@ -318,7 +311,7 @@ library SafePermit {
                 mstore(add(0x40, ptr), keccak256(add(0x40, ptr), 0x80))
                 signingHash := keccak256(add(0x1e, ptr), 0x42)
             }
-            _checkSignature(signingHash, owner, v, r, s);
+            _checkSignature(signingHash, owner, vs, r);
         }
     }
 }