+++ test/integration/Permit.sol
@@ -31,7 +31,7 @@ contract PermitTest is SettlerBasePairTest {
     }
 
     function _testName() internal pure override returns (string memory) {
-        return "ERC2612Funding";
+        return "transfer-from-with-permit";
     }
 
     function fromToken() internal pure override returns (IERC20) {}
@@ -40,6 +40,10 @@ contract PermitTest is SettlerBasePairTest {
 
     function amount() internal pure override returns (uint256) {
         return 1000e6;
+    }
+
+    function _vs(uint256 v, bytes32 s) internal view returns (bytes32 vs) {
+        return bytes32(v - 27) << 255 | s;
     }
 
     function _signERC2612Permit(address owner, address spender, uint256 value, uint256 deadline, uint256 privateKey)
@@ -102,7 +106,7 @@ contract PermitTest is SettlerBasePairTest {
         (uint8 v, bytes32 r, bytes32 s) = _signERC2612Permit(sender, address(allowanceHolder), amount(), deadline, pk);
 
         bytes memory permitData =
-            abi.encodePacked(Permit.PermitType.ERC2612, abi.encode(sender, amount(), deadline, v, r, s));
+            abi.encodePacked(Permit.PermitType.ERC2612, abi.encode(amount(), deadline, r, _vs(v, s)));
 
         bytes[] memory actions = ActionDataBuilder.build(
             abi.encodeCall(
@@ -114,6 +118,7 @@ contract PermitTest is SettlerBasePairTest {
         uint256 snapshot = vm.snapshot();
 
         vm.prank(sender);
+        snapStartName("ERC2612");
         allowanceHolder.exec(
             address(settler),
             address(USDC),
@@ -130,6 +135,7 @@ contract PermitTest is SettlerBasePairTest {
                 )
             )
         );
+        snapEnd();
 
         assertEq(USDC.balanceOf(address(this)), amount(), "Transfer failed");
         assertEq(USDC.balanceOf(sender), 0, "Sender should have 0 balance");
@@ -191,7 +197,7 @@ contract PermitTest is SettlerBasePairTest {
         (uint8 v, bytes32 r, bytes32 s) = _signDAIPermit(sender, address(allowanceHolder), nonce, expiry, true, pk);
 
         bytes memory permitData =
-            abi.encodePacked(Permit.PermitType.DAIPermit, abi.encode(sender, nonce, expiry, true, v, r, s));
+            abi.encodePacked(Permit.PermitType.DAIPermit, abi.encode(nonce, expiry, true, r, _vs(v, s)));
 
         bytes[] memory actions = ActionDataBuilder.build(
             abi.encodeCall(
@@ -203,6 +209,7 @@ contract PermitTest is SettlerBasePairTest {
         uint256 snapshot = vm.snapshot();
 
         vm.prank(sender);
+        snapStartName("DAIPermit");
         allowanceHolder.exec(
             address(settler),
             address(DAI),
@@ -219,7 +226,8 @@ contract PermitTest is SettlerBasePairTest {
                 )
             )
         );
-
+        snapEnd();
+        
         assertEq(DAI.balanceOf(address(this)), amount(), "Transfer failed");
         assertEq(DAI.balanceOf(sender), 0, "Sender should have 0 balance");
 
@@ -280,7 +288,7 @@ contract PermitTest is SettlerBasePairTest {
         (uint8 v, bytes32 r, bytes32 s) = _signNativeMetaTransaction(sender, address(allowanceHolder), amount(), pk);
 
         bytes memory permitData =
-            abi.encodePacked(Permit.PermitType.NativeMetaTransaction, abi.encode(sender, amount(), v, r, s));
+            abi.encodePacked(Permit.PermitType.NativeMetaTransaction, abi.encode(amount(), r, _vs(v, s)));
 
         bytes[] memory actions = ActionDataBuilder.build(
             abi.encodeCall(
@@ -292,6 +300,7 @@ contract PermitTest is SettlerBasePairTest {
         uint256 snapshot = vm.snapshot();
 
         vm.prank(sender);
+        snapStartName("NativeMetaTransaction");
         allowanceHolder.exec(
             address(settler),
             address(ROUTE),
@@ -308,6 +317,7 @@ contract PermitTest is SettlerBasePairTest {
                 )
             )
         );
+        snapEnd();
 
         assertEq(ROUTE.balanceOf(address(this)), amount(), "Transfer failed");
         assertEq(ROUTE.balanceOf(sender), 0, "Sender should have 0 balance");